+++ management/commands/load_geometadata_demo.py
@@ -45,6 +45,7 @@ DEMO_PLUGIN_SETTINGS = {
     "embed_geo_meta": "on",
     "embed_schema_spatial": "on",
     "embed_geojson_link": "on",
+    "embed_iso19139": "on",
 }